Tweak load/save API
- Add Font::load and Font::from_bytes methods
- Change Font::save to take a path instead of a handle

This is hopefully a simpler API, not requiring the user to mess
around with files directly.

I'm not sure if there are uses for font::load that are not covered
by the new methods, so I have left it in place but marked it as
deprecated; if it should be removed I'm happy to make that change.