Fixup tag! macro export situation
The tag! macro generates code that looks approximately like:
`crate::types::Tag::unsafe_constructor(input_bytes)`. The tricky part
here is the 'crate' bit. What should go here?

The problem has to do with where the macro is exported. Currently it is
exported from both the `otspec` and the `fonttools` crates. This creates
a problem: if the macro generates `otspec::types::Tag(..)` then it won't
compile when used from `fonttools` if `otspec` is not in scope; and if
it generates `fonttools::types::Tag(..)` then it won't work when used
from otspec if fonttools is not in scope.

I've gone with a simple solution, which is to *only* export tag! from
fonttools, and to generate 'fonttools::types::Tag(..)'. There are fancier
solutions available (such as https://crates.io/crates/proc-macro-crate) that
we can turn to if this ends up being a problem, but I do not expect in practice
that many folks are going to use otspec outside of fonttools.